+++ stacks/src/test/x400p-ss7/test-x100p.c
@@ -7536,7 +7536,7 @@ test_9_3(void)
 		return INCONCLUSIVE;
 	if (msu_len > 12)
 		msu_len = 12;
-	printf("(N1=%ld, N2=%ld, n=%d, l=%d)\n", iutconf.sl.N1, iutconf.sl.N2, n, msu_len);
+	printf("(N1=%ld, N2=%ld, n=%d, l=%d)\n", (long) iutconf.sl.N1, (long) iutconf.sl.N2, n, msu_len);
 	fflush(stdout);
 	for (;;) {
 		switch (state) {
@@ -7620,7 +7620,7 @@ test_9_4(void)
 	n = iutconf.sl.N2 / (msu_len + h) + 1;
 	if (msu_len > iutconf.sdt.m)
 		return INCONCLUSIVE;
-	printf("(N1=%ld, N2=%ld, n=%d, l=%d)\n", iutconf.sl.N1, iutconf.sl.N2, n, msu_len);
+	printf("(N1=%ld, N2=%ld, n=%d, l=%d)\n", (long) iutconf.sl.N1, (long) iutconf.sl.N2, n, msu_len);
 	fflush(stdout);
 	for (;;) {
 		switch (state) {
@@ -7710,7 +7710,7 @@ test_9_5(void)
 	}
 	if (msu_len > 12)
 		msu_len = 12;
-	printf("(N1=%ld, N2=%ld, n=%d, l=%d)\n", iutconf.sl.N1, iutconf.sl.N2, n, msu_len);
+	printf("(N1=%ld, N2=%ld, n=%d, l=%d)\n", (long) iutconf.sl.N1, (long) iutconf.sl.N2, n, msu_len);
 	fflush(stdout);
 	for (;;) {
 		switch (state) {
@@ -7802,7 +7802,7 @@ test_9_6(void)
 	}
 	if (msu_len > 12)
 		msu_len = 12;
-	printf("(N1=%ld, N2=%ld, n=%d, l=%d)\n", iutconf.sl.N1, iutconf.sl.N2, n, msu_len);
+	printf("(N1=%ld, N2=%ld, n=%d, l=%d)\n", (long) iutconf.sl.N1, (long) iutconf.sl.N2, n, msu_len);
 	fflush(stdout);
 	for (;;) {
 		switch (state) {
